third_party/WebKit/LayoutTests/platform/mac/external/wpt/IndexedDB/nested-cloning-large-multiple-expected.txt

This is a testharness.js-based test.
PASS multiple requests of objects with blobs and large typed arrays 
FAIL multiple requests of objects with blobs and large typed arrays with key generator assert_equals: Expected success event, but got error event instead expected "success" but got "error"
Harness: the test ran to completion.
